requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.25.0
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
python-telegram-bot==21.5
tweepy==4.14.0  # para twitter/X
transformers==4.44.2
torch==2.4.1
streamlit==1.38.0
python-dotenv==1.0.1
schedule==1.2.2
pytest==8.3.2
pydrive2==1.19.4  # google drive

flake8==7.0.0
pytest-cov==5.0.0